pyproject.toml
[tool.poetry]
name = "sign-system"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-login = "^0.6.3"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.32"
flask-sqlalchemy = "^3.1.1"
cryptography = "^43.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
